Update the test project
The test project is updated to .NET 9, including an update of all the nuget packages.

The update to NUnit 4 requires the use of the fluent style of tests. The test files are now updated to use this style.

The test files are now updated to use file scoped namespaces. Some file were missing a namespace, now they have a namespace, and some classes were `internal`, now they are `public` like the other test classes.